basic admin module + backend splitted on profiles and postgres added

+++ store-backend/src/main/java/ru/voroby/storebackend/repository/ProductDAO.java
@@ -3,7 +3,7 @@ package ru.voroby.storebackend.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.voroby.storebackend.model.Product;
 
-public interface ProductDAO extends CrudRepository<Product, Long> {
+public interface ProductDAO extends CrudRepository<Product, Integer> {
 
   Product findByNameAndCategory(String name, String category);
